+++ tests/test_builtins.rs
@@ -2,6 +2,6 @@ use bindgen;
 
 #[test]
 fn test_builtin_va_list() {
-	let bindings = bindgen::Builder::new().header("tests/headers/builtin_va_list.h").builtins().generate().unwrap().to_string();
+	let bindings = bindgen::Builder::new("tests/headers/builtin_va_list.h").builtins().generate().unwrap().to_string();
     assert!(bindings.contains("__builtin_va_list"));
 }